Report w Images, Deleted Case Notes

+++ resources/views/case_note_actions/index.blade.php
@@ -20,27 +20,29 @@
         <div class='table-responsive padding-top'>
             <table class="table table-hover">
                 <tr>
-                    <th>Case Id</th>
+                    <th>Case ID</th>
                     <th>Date</th>
                     <th>Time</th>
                     <th>Title</th>
                     <th>Description</th>
-                    <th>Image</th>
+                    <th class="max-width-med">Image</th>
                     <th>Reporting Guard</th>
                     <th>Manage</th>
                 </tr>
 
                 @foreach($cases as $index => $note)
-                    <tbody class="group-list">
-                    <tr>
-                        <td class="report-title" colspan="4">{{$index}}</td>
+
+                    <tr class="report-title-bg">
+                        <td colspan="4">{{$index}}</td>
                         <td></td>
                         <td></td>
                         <td></td>
                         <td></td>
                     </tr>
                     @foreach ($cases->get($index) as $item)
-                        <tr>
+                    @if(($item->title != "Nothing to Report")&&($item->deleted_at == null))
+                        <tbody class="alt-cols">
+                            <tr>
                             <td class="padding-left max-width">{{$item->case_id}}</td>
 
                             @if(isset($item->date))
@@ -52,11 +54,12 @@
                             @if(isset($item->time))
                                 <td class="min-width">{{$item->time}}</td>
                             @else
-                                <td>GeoData n/a</td>
+                                <td>-</td>
                             @endif
 
-                            <td>{{$item->title}}</td>
-                            <td class="desc-max-width">{{$item->description}}</td>
+                            <td class="max-width-big">{{$item->title}}</td>
+
+                            <td class="desc-max-width min-width">{{$item->description}}</td>
 
                             @if($item->hasImg == "Y")
                                 @if(isset($item->files))
@@ -64,11 +67,14 @@
                                     @if(sizeof($item->files) > 0)
                                         {{--first photo in array--}}
                                         {{--<td><a href="{{$item->urls[0]}}" target="_blank">Download {{$item->files[0]}}</a></td>--}}
-                                        <td><img src="{{$item->urls[0]}}" height="25px" width="25px"/></td>
+                                        <td><a href="{{$item->urls[0]}}" target="_blank">
+                                                <img src="{{$item->urls[0]}}" height="75px" width="75px"/>
+                                            </a>
+                                        </td>
                                     @else
                                         {{--v2 uploads--}}
                                         {{--todo: remove once not using v2 mobile anymore--}}
-                                        <td><a href="{{$item->url}}" target="_blank">WIP</a></td>
+                                        <td>Mobile App V2 Image</td>
                                         {{--<td><img src="{{$item->url}}"/></td>--}}
 
                                     @endif
@@ -84,15 +90,24 @@
 
                             <td>{{$item->first_name}} {{$item->last_name}}</td>
 
-                            @if($item->title != "Nothing to Report")
-                                <td><a href="/case-notes/{{$item->id}}/edit" class="edit-links"><i class="fa fa-edit"></i></a>
-                                    <a href="/confirm-delete/{{$item->id}}/{{$url}}"
-                                       style="color: #990000;"><i class="fa fa-trash-o icon-padding"></i></a>
-                                </td>
+                            @if($item->deleted_at == null)
+
+                                @if($item->title != "Nothing to Report")
+                                    <td><a href="/case-notes/{{$item->id}}/edit" class="edit-links"><i class="fa fa-edit"></i></a>
+                                        <a href="/confirm-delete/{{$item->id}}/{{$url}}"
+                                           style="color: #990000;"><i class="fa fa-trash-o icon-padding"></i></a>
+                                    </td>
+                                @else
+                                    <td>
+                                        <a href="/confirm-delete/{{$item->id}}/{{$url}}"
+                                           style="color: #990000;"><i class="fa fa-trash-o"></i></a>
+                                    </td>
+                                @endif
                             @else
                                 <td>
-                                    <a href="/confirm-delete/{{$item->id}}/{{$url}}"
-                                       style="color: #990000;"><i class="fa fa-trash-o"></i></a>
+                                    <a href="/delete/{{$urlCancel}}/{{$item->case_note_id}}/{{$report->id}}" style="color: #990000;" disabled>
+                                        <i class="fa fa-trash-o"></i>
+                                    </a>
                                 </td>
                             @endif
                         </tr>
@@ -107,17 +122,16 @@
                                         <td></td>
                                         <td></td>
                                         <td></td>
-                                        <td><img src="{{$item->urls[$i]}}" height="25px" width="25px"/></td>
-                                        {{--<td><a href="{{$item->urls[$i]}}" target="_blank">Download Image {{$i + 1}}</a></td>--}}
+                                        <td><img src="{{$item->urls[$i]}}" height="75px" width="75px" /></td>
                                         <td></td>
                                         <td></td>
                                     </tr>
                                 @endfor
                             @endif
                         @endif
-
+                        </tbody>
+                    @endif
                     @endforeach
-                    </tbody>
                 @endforeach
 
             </table>
